bugs
more int to float

+++ script.js
@@ -105,13 +105,13 @@ function Save(){
 	var bossHP = parseInt(document.getElementById('bossHP').value);
 
 	e = document.getElementById("adv1");
-	var adv1 = parseInt(e.options[e.selectedIndex].text);
+	var adv1 = parseFloat(e.options[e.selectedIndex].text);
 	e = document.getElementById("adv2");
-	var adv2 = parseInt(e.options[e.selectedIndex].text);
+	var adv2 = parseFloat(e.options[e.selectedIndex].text);
 	e = document.getElementById("adv3");
-	var adv3 = parseInt(e.options[e.selectedIndex].text);
+	var adv3 = parseFloat(e.options[e.selectedIndex].text);
 	e = document.getElementById("advBoss");
-	var advBoss = parseInt(e.options[e.selectedIndex].text);
+	var advBoss = parseFloat(e.options[e.selectedIndex].text);
 
 	var stats = [{
 		"atk1":atk1,
